+++ src/Medidata.MAuth.Core/MAuthSigningHandler.cs
@@ -12,11 +12,10 @@ namespace Medidata.MAuth.Core
     /// </summary>
     public class MAuthSigningHandler: DelegatingHandler
     {
-        private readonly MAuthSigningOptions options;
-        private IMAuthCore mAuthCore;
+        private readonly MAuthSigningOptions _options;
 
         /// <summary>Gets the Uuid of the client application.</summary>
-        public Guid ClientAppUuid => options.ApplicationUuid;
+        public Guid ClientAppUuid => _options.ApplicationUuid;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="MAuthSigningHandler"/> class with the provided
@@ -25,7 +24,7 @@ namespace Medidata.MAuth.Core
         /// <param name="options">The options for this message handler.</param>
         public MAuthSigningHandler(MAuthSigningOptions options)
         {
-            this.options = options;
+            _options = options;
         }
 
         /// <summary>
@@ -38,7 +37,7 @@ namespace Medidata.MAuth.Core
         /// </param>
         public MAuthSigningHandler(MAuthSigningOptions options, HttpMessageHandler innerHandler): base(innerHandler)
         {
-            this.options = options;
+            _options = options;
         }
 
         /// <summary>
@@ -55,16 +54,14 @@ namespace Medidata.MAuth.Core
             if (InnerHandler == null)
                 InnerHandler = new HttpClientHandler();
 
-            if (options.DisableV1 == false) // default
+            foreach (MAuthVersion version in Enum.GetValues(typeof(MAuthVersion)))
             {
-                // Add headers for V1 protocol as well
-                var mAuthCoreV1 = MAuthCoreFactory.Instantiate(MAuthVersion.MWS);
-                request = await mAuthCoreV1.Sign(request, options).ConfigureAwait(false);
+                if (_options.SignVersions.HasFlag(version))
+                {
+                    var mAuthCore = MAuthCoreFactory.Instantiate(version);
+                    request = await mAuthCore.Sign(request, _options).ConfigureAwait(false);
+                }
             }
-
-            // Add headers for V2 protocol
-            mAuthCore = MAuthCoreFactory.Instantiate(MAuthVersion.MWSV2);
-            request = await mAuthCore.Sign(request, options).ConfigureAwait(false);
 
             return await base
                 .SendAsync(request, cancellationToken)