+++ tests/Medidata.MAuth.Tests/MAuthSigningHandlerTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Net.Http;
+﻿using System.Collections.Generic;
+using System.Net.Http;
 using System.Threading.Tasks;
 using Medidata.MAuth.Core;
 using Medidata.MAuth.Core.Models;
@@ -14,12 +15,71 @@ namespace Medidata.MAuth.Tests
         [InlineData("DELETE")]
         [InlineData("POST")]
         [InlineData("PUT")]
-        public static async Task SendAsync_WithDefault_WillSignProperly_BothMWSAndMWSV2(string method)
+        public static async Task SendAsync_WithNoSignVersions_WillSignWithOnlyMWS(string method)
         {
             // Arrange
             var testData = await method.FromResource();
             var actual = new AssertSigningHandler();
+            var version = MAuthVersion.MWS;
             var signingHandler = new MAuthSigningHandler(TestExtensions.ClientOptions(testData.SignedTime), actual);
+
+            // Act
+            using (var client = new HttpClient(signingHandler))
+            {
+                await client.SendAsync(testData.ToHttpRequestMessage(version));
+            }
+
+            // Assert
+            Assert.Null(actual.MAuthHeaderV2);
+            Assert.Null(actual.MAuthTimeHeaderV2);
+
+            Assert.Equal(testData.MAuthHeader, actual.MAuthHeader);
+            Assert.Equal(testData.SignedTime, long.Parse(actual.MAuthTimeHeader).FromUnixTimeSeconds());
+        }
+
+        [Theory]
+        [InlineData("GET")]
+        [InlineData("DELETE")]
+        [InlineData("POST")]
+        [InlineData("PUT")]
+        public static async Task SendAsync_WithSignVersionMWS_WillSignWithOnlyMWS(string method)
+        {
+            // Arrange
+            var testData = await method.FromResourceV2();
+            var actual = new AssertSigningHandler();
+            var clientOptions = TestExtensions.ClientOptions(testData.SignedTime);
+            var version = MAuthVersion.MWS;
+            clientOptions.SignVersions = version;
+            var signingHandler = new MAuthSigningHandler(clientOptions, actual);
+
+            // Act
+            using (var client = new HttpClient(signingHandler))
+            {
+                await client.SendAsync(testData.ToHttpRequestMessage(version));
+            }
+
+            // Assert
+            Assert.Equal(testData.MAuthHeader, actual.MAuthHeader);
+            Assert.Equal(testData.SignedTime, long.Parse(actual.MAuthTimeHeader).FromUnixTimeSeconds());
+
+            Assert.Null(actual.MAuthHeaderV2);
+            Assert.Null(actual.MAuthTimeHeaderV2);
+        }
+
+        [Theory]
+        [InlineData("GET")]
+        [InlineData("DELETE")]
+        [InlineData("POST")]
+        [InlineData("PUT")]
+        public static async Task SendAsync_WithSignVersionsMWSAndMWSV2_WillSignWithBothMWSAndMWSV2(string method)
+        {
+            // Arrange
+            var testData = await method.FromResourceV2();
+            var actual = new AssertSigningHandler();
+            var clientOptions = TestExtensions.ClientOptions(testData.SignedTime);
+            clientOptions.SignVersions = MAuthVersion.MWS | MAuthVersion.MWSV2;
+
+            var signingHandler = new MAuthSigningHandler(clientOptions, actual);
 
             // Act
             using (var client = new HttpClient(signingHandler))
@@ -31,33 +91,6 @@ namespace Medidata.MAuth.Tests
             Assert.Equal(testData.MAuthHeader, actual.MAuthHeader);
             Assert.Equal(testData.SignedTime, long.Parse(actual.MAuthTimeHeader).FromUnixTimeSeconds());
 
-            Assert.Equal(testData.MAuthHeaderV2, actual.MAuthHeaderV2);
-            Assert.Equal(testData.SignedTime, long.Parse(actual.MAuthTimeHeader).FromUnixTimeSeconds());
-        }
-
-        [Theory]
-        [InlineData("GET")]
-        [InlineData("DELETE")]
-        [InlineData("POST")]
-        [InlineData("PUT")]
-        public static async Task SendAsync_WithDisableV1_WillSignProperlyWithMWSV2(string method)
-        {
-            // Arrange
-            var testData = await method.FromResourceV2();
-            var actual = new AssertSigningHandler();
-            var version = MAuthVersion.MWSV2;
-            var clientOptions = TestExtensions.ClientOptions(testData.SignedTime);
-            clientOptions.DisableV1 = true;
-            var signingHandler = new MAuthSigningHandler(clientOptions, actual);
-
-
-            // Act
-            using (var client = new HttpClient(signingHandler))
-            {
-                await client.SendAsync(testData.ToHttpRequestMessage(version));
-            }
-
-            // Assert
             Assert.Equal(testData.MAuthHeaderV2, actual.MAuthHeaderV2);
             Assert.Equal(testData.SignedTime, long.Parse(actual.MAuthTimeHeaderV2).FromUnixTimeSeconds());
         }